refactor: cursor.cursor_buf -> cursor.buf
